githubAPI/requirements.txt
curl_cffi==0.7.1
python-dotenv==1.0.1
pytest==8.3.3